refactor(project): Remove some unnecessary code and reformat code
