--- api/src/main/java/com/mtech/vmcs/exception/ControllerAdvisor.java
@@ -1,6 +1,0 @@
-package com.mtech.vmcs.exception;
-
-import org.springframework.web.bind.annotation.ControllerAdvice;
-
-@ControllerAdvice
-public class ControllerAdvisor {}
