--- api/src/main/java/com/mtech/vmcs/model/enums/ExceptionEnum.java
@@ -1,3 +1,0 @@
-package com.mtech.vmcs.model.enums;
-
-public enum ExceptionEnum {}
