+++ api/src/main/java/com/mtech/vmcs/service/DrinkService.java
@@ -6,8 +6,6 @@ import java.util.List;
 
 public interface DrinkService {
 
-  List<Drink> getDrinksByName(String name);
-
   List<Drink> getAllDrinks();
 
   void createDrinks(List<Drink> drinks);
