+++ api/src/main/java/com/mtech/vmcs/util/MementoStack.java
@@ -1,10 +1,10 @@
-package com.mtech.vmcs.utill;
+package com.mtech.vmcs.util;
 
 import java.util.Stack;
 
 public class MementoStack<T> {
 
-  Stack<T> stack = new Stack<>();
+  private final Stack<T> stack = new Stack<>();
 
   public void push(T state) {
     stack.push(state);
